refactor: separate errors for `add_cert` and `add_vote` (#32)
I was expecting this change to be more impactful than it turned out to
be. I still think this is a good change to have long term but the status
quo is perfectly fine for now.

We define different errors for the two functions so that they can evolve
independently without impacting the callers of each others.

With this change, we do not need to track what error types can be
returned in comments as the expectation should be that all variants
should be returnable.

---------

Co-authored-by: Quentin Kniep <quentin@anza.xyz>