refactor: Introduce a Slot type (#41)
Before, we stored the slot as a simple `u64` which is dangerous as it
can be mixed up with other unrelated types. This PR defines `Slot` as a
`struct(u64)` to provide us with some type safety.

In general, I am not happy with the use of `SLOTS_PER_WINDOW` in so many
tests. But I am not sure if it makes sense to update them at this point
in time. We can discuss if we want to refactor them in this PR or a
subsequent PR.

---------

Co-authored-by: Quentin Kniep <quentin@anza.xyz>