fix: Reed-Solomon padding (#69)
Reed-Solomon shredding and deshredding currently breaks for various
slice data sizes (including empty slices and uneven number of bytes per
shred). This was not caught by any tests.

Added some tests with variously sized slices to catch problems like
these. Implemented padding to make those tests pass.

If I understand the requirements of the `reed-solomon-simd` crate
correctly, the minimum padding we need to do is to pad payload data to a
multiple of `2 * DATA_SLICES`. The alternative would be to simply always
pad to the full `MAX_DATA_PER_SLICE`. We also have to account for
storing the integer indicating the number of bytes of padding that was
added.

---------

Co-authored-by: Akhilesh Singhania <akhi3030@gmail.com>