fix: repair missing ancestors (#78)
This addresses #31.

The goal was to repair any missing ancestors of a block that is finalized. Instead of repairing only ancestors of finalized blocks, we immediately start repairing the ancestors of any block we repair. In the common case (where there are no equivocations) this is even better because we can catch up faster. In the case with equivocation, this could lead to us repairing blocks that we will not need in the end.