feat: ensure we do not store shreds and slices with invalid slice indexes (#101)
Closes: https://github.com/qkniep/alpenglow/issues/100.

Introduces a `SliceIndex` type which is guaranteed to be in the range
`[0, MAX_SLICES_PER_BLOCK]`. And updates various datastructures to use
`SliceIndex` instead of `usize`. Now, we will essentially drop invalid
shreds right when we receive them at the network layer and we are
guaranteed that various structs like blockstore, etc. are not storing
invalid shreds and slices.

---------

Co-authored-by: Quentin Kniep <quentin@anza.xyz>
Co-authored-by: Quentin Kniep <hello@quentinkniep.com>