test: adds another simple test for `wait_for_first_slot` (#102)
I am a bit perplexed that this test is passing.

Relevant code:
[here](https://github.com/qkniep/alpenglow/blob/6e3fa7ba4dab97e250744f7db81ef17c5017b8ec/src/consensus.rs#L136).

Because we are using `tokio::select`, I would've expected to see some
calls into blockstore as well which will cause the test to panic because
I have not specified any expectations on blockstore. I ran the test
several times over a period of 20mins and it always passed.

So I feel like my understanding of tokio::select is lacking.

My rough intuition is that select has been given two futures, the
receiver and the task. Select is somehow able to figure out that the
receiver is ready to produce an output so it doesn't even bother
inspecting the task. Note that changing the order of the futures in
production code does not change impact the test (which is a relief).

@qkniep: let me know if you think such tests are not worth having. I
feel like they are bordering a bit on change detector type of tests.

---------

Co-authored-by: Quentin Kniep <hello@quentinkniep.com>
Co-authored-by: Quentin Kniep <quentin@anza.xyz>