fix: shred indexing and more tests to `blockstore` module (#74)
This is part of #71.

While adding tests to the `blockstore` module, I also rediscovered and
had to fix an old problem, that needed to be fixed for the new tests to
pass: The `index_in_slice` for shreds was previously going e.g. 0..32
for data shreds and 0..32 for coding shreds again, instead coding shred
indices should be offset (e.g. 0..32 for data shreds and 32..64 for
coding shreds). This way this index uniquely refers to a shred, can be
used directly in repair, everywhere in blockstore and for the Merkle
trees. Only the underlying reed-solomon implementation still requires
the old indexing, so when interfacing with it after shredding and before
deshredding we have to translate the indexing.