bug: emit `ParentReady` for ancestor of finalized block  (#99)
When seeing a new finalized block, Pool should check whether we issued a
`ParentReady` for this window already. If not a `ParentReady` should be
issued for the first slot of the window.

This ensures that we correctly enter the window when catching up based
on finalized blocks, setting the timeouts and voting notar or skip for
the slots. Especially during standstill this is critical. Without this,
if enough validators are in this state, the network may not ever recover
from standstill.

Closes #97.

---------

Co-authored-by: Akhilesh Singhania <akhi3030@gmail.com>