nit: rename `write_bytes` to `to_slice`, pass error, refactor (#117)
Serialization function with reusable buffer was renamed from
`write_bytes()` to `to_slice()` to be aligned with the analogous
`to_bytes()`

Instead of panicking if the buffer is to small for the message being
serialized it now passes the error up to the caller, only panicking on
unexpected bincode errors.

Refactored out the bincode config being used into a `const` while I am
at it.

See also:
https://github.com/qkniep/alpenglow/pull/112#discussion_r2296532037