fix: optimistic block production leading to skips (#120)
Optimistic block production was blocking on producing a slice with
timeout duration `Duration::MAX` before the `ParentReady` event has been
seen. If this slice never completed, e.g. because we don't see any
transaction, we block forever and never start the blocktime timeout. The
way I fixed this now, is by limiting the time per slice before
`ParentReady` was seen to a maximum of `DELTA_BLOCK / 2`.

Closes #119.