feat: a simple workload generator (#118)
It is it's own binary called `workload_generator` and can generate and
send random transaction of a given size and at a given TPS to the
validator's transaction receiver port (all provided as CLI arguments).

Some initial runs on my MacBook show that it currently maxes out at
~320k TPS (of 20 bytes each, ~50 Mbit/s) or ~200k TPS (of 512 bytes
each, ~800 Mbit/s).

Closes #47.
Part of #13.