feat: perform actual block repair (#106)
Closes #12.

So far repair only requests the parent information. We then ingest the
parent information into `Pool`. This is enough for consensus to
continue. However, it bypasses the main point of repairing the actual
block. Also, currently the `Parent` responses do not contain any proof
that the parent is the actual parent of the requested block. In this PR
we replace this with actual repair of block, where each response
contains a proof.

---------

Co-authored-by: Akhilesh Singhania <akhi3030@gmail.com>