feat: faster Merkle tree construction, more max nodes, handle edge cases (#150)
Handles some open tasks in the `crypto::merkle` module. Some
out-of-bounds accesses have assertions now. Also increased max number of
leaves from 2^24 to 2^32, providing the efficient function used to
calculate those.

Maybe most important, we now use the pre-calculated `EMPTY_ROOTS` to
speed up the constructor (up to a factor of two if the second half of
the tree is almost empty).

Closes #148.