fix: finalization slot/block, naming around certs/finalization, adapt tests (#158)
Previously, the basically meaningless notion of a finalized slot without
a block was wrongfully exposed and used (i.e. the case when we have a
finalization cert but don't yet have the notarization cert for a block).
This is now fixed, only exposing the notion of notarized block as having
fast finalization or final+notar certs outside of `FinalityTracker`. To
go along with this, some of the naming around finalization and
certificates has been adapted to be more accurate and less error prone.