refactor: clean up `NetworkError` (#168)
Splits `NetworkError` into `NetworkReceiveError` and `NetworkSendError`.

This eliminates various places where we check for error variants that
could not happen and also allows us to reduce the number of error
variants overall. And in one scenario, in repair, eliminates a panic
that can happen if we receive a message cannot be deserialized.