nit: change an unexpected error condition to panic (#175)
`NetworkMessage::to_slice()` is only used in testing. Even when we start
using it in production, I do not imagine us writing code to handle the
situation that the buffer was too small and calling the function again
with a larger buffer so I think it is cleaner for the function to panic
if the buffer is not big enough.