refactor: constrain what types of messages `Network` can handle (#183)
Also see https://github.com/qkniep/alpenglow/pull/177 for more
discussions.

High level idea is that consumers of `Network::receive()` have to
sometimes handle message types that they are not interested in. They
generally handle this by ignoring them and continuing to loop. We
already have a similar loop in the structs that implement `Network` so
the extra looping is unergonomic.

This patch adds constraints to `Network` that allows one to specify what
types of messages they are willing to send and receive. This allows the
consumers to specify what types of messages they are interested in and
also eliminate the extra looping.

Potential future work being tracked in
https://github.com/qkniep/alpenglow/issues/187.

---------

Co-authored-by: Quentin Kniep <hello@quentinkniep.com>