bug: fix some broadcast test failures (#215)
In https://github.com/qkniep/alpenglow/pull/213, we observed some test
failures. Debugging it, I believe the following is happening:

- The sender is broadcasting and then exiting without receiving the
messages that it broadcast
- if we send the messages too quickly as is happening in #213, the
sender is exiting before simulated/core is able to deliver the msgs to
it.
- the sender exits and then the core tries to send it message, finds
that the channel is closed, and panics
- now the other receivers get stuck as they have not received all of the
messages they wanted

The proposed fix is for the sender to receive messages that it broadcast
before exiting and for the the core to not panic if it fails deliver
messages.

---------

Co-authored-by: Quentin Kniep <quentin@anza.xyz>