refactor: do not construct potentially invalid `Shred` in `reed_solomon_shred()` (#221)
`reed_solomon_shred()` is returning `DataShred` and `CodingShred` and
assigning `index_in_shred` in them. It does not have enough context to
assign the right index in the shreds. The actual index depends on the
actual type of shredder that we are using.

See https://github.com/qkniep/alpenglow/pull/190/files#r2376067497 for
more context.

This PR updates `reed_solomon_shred()` to return just the shredded
payload instead. Then the individual shredders construct the actual
`Shred`s with valid indexes.