refactor: store `BlockData::shreds` as arrays (#190)
Closes: #182 

This PR ended up being much bigger than I originally hoped.

The key change that triggered the whole PR is to store
`BlockData::shreds` in an array of size `TOTAL_SHREDS`.

In order to make sure that we can index safely into the array and not
have to handle out of bounds errors, I then introduced `ShredIndex` type
to bound how we index into the array and used it in appropriate places.

This caused some places where we construct `ShredPayload` to use the new
`ShredIndex` and either return errors or use `unwrap`s if we can
guarantee that the number of shreds will be bound by `TOTAL_SHREDS`.
This requirement then caused me to change various interfaces in
`shredders.rs` and `reed_solomon.rs` to use arrays instead of `Vec`.

---------

Co-authored-by: Quentin Kniep <hello@quentinkniep.com>
Co-authored-by: Quentin Kniep <quentin@anza.xyz>