refactor: actual newtypes for Merkle proofs/roots, including `BlockHash` type (#253)
Introduces newtypes `SliceRoot` and `SliceProof` for `SliceMerkleTree`
and `DoubleMerkleRoot` and `DoubleMerkleProof` for `DoubleMerkleTree`.
Also, introduces a type alias `BlockHash` for `DoubleMerkleRoot`. These
types should now be used almost everywhere where previously `Hash` was
used.

The single remaining use case of `Hash` that could be replaced is within
repair, where a hash is used as nonce identifier for outstanding
requests.

Builds on #252.

Closes #247, #248, #249, #250.