feat: change padding scheme to one that supports >=128 data shreds (#259)
So far, the padding scheme was including number of padding bytes in each
of the last bytes, i.e. padding data to a multiple of 4 bytes might look
like this:
`DD DD DD DD | DD 03 03 03`
One downside to this scheme is that the highest number of padding bytes
it supports is 255. In our cases, where we want to pad to a multiple of
`2 * DATA_SHREDS` this introduces the limit `DATA_SHREDS <= 127`.

The new padding scheme pads with one `0x80` byte and zero or more `0x00`
bytes (this is equivalent to adding a single `1` bit and then `0` bits
as necessary), for example:
`DD DD DD DD | DD 80 00 00`
This scheme works for any number of padding bytes (just as in the old
scheme we always add at least one byte of padding to ensure we are
always able to differentiate data from padding).