nit: store notar fallback votes in a `BTreeMap` (#260)
Before we were storing them in a `Vec` and this PR changes it to a
`BTreeMap`. This should not change the performance at all as the size of
the container should never get too large but helps the reader in
understanding the code better and also can help with accidental bugs.