fix: reserve network resource after calculating start time, not before (#270)
When writing docs for #269, I noticed a bug in the `broadcast()`
function of the `discrete_event_simulation` module. The network resource
was reserved for the entire duration of the broadcast before calculating
the start time for sending. This results in a wrong (delayed) start time
for the broadcast.

Instead, the network resource is now correctly reserved after
calculating the start time for sending.