perf: introduce `ShredderPool` (#287)
Even though #261 changed the `reed_solomon` API to allow for reuse of
the shredder, blockstore is currently creating new shredder for each
slice.

Here we actually make the blockstore keep a `ShredderPool`. So far this
only has one shredder, because adding shreds to the blockstore is
single-threaded anyways (always takes a `&mut` of the blockstore). In
the future this could hold up to one shredder per thread used in the
blockstore. Also, it might make sense to use the `ShredderPool` in other
places, e.g. the block producer.

Follow-up to #261.
Supersedes #286.

---------

Co-authored-by: Akhilesh Singhania <akhi3030@gmail.com>