chore: change wincode dependency back to crates.io (#289)
Now that v0.2 of the `wincode` crate is released, which includes my PR
(https://github.com/anza-xyz/wincode/pull/15) that we depend on, we can
change the dependency back from the specific GitHub commit to a stable
crates.io version.

With the v0.2 release there were also some changes to the API that we
needed to adapt to.

Closes #281.